temp: refactoring download and upload features
