+++ src/app/openai/page.tsx
@@ -3,7 +3,7 @@ import { OpenAIClientComponent } from "./client";
 import { OPENAI, UTILS } from "~/lib/infrastructure/server/config/ioc/server-ioc-symbols";
 import type {  File } from "~/lib/core/entity/file";
 import type { Logger } from "pino";
-import type SourceDataRepositoryOutputPort from "~/lib/core/ports/secondary/source-data-repository-output-port";
+import type SourceDataGatewayOutputPort from "~/lib/core/ports/secondary/source-data-gateway-output-port";
 import fs from "fs";
 import type RemoteStorageElementOutputPort from "~/lib/core/ports/secondary/remote-storage-element-output-port";
 
@@ -25,7 +25,7 @@ const uploadFile = async (file: File) => {
     // await OpenAIRSE.uploadFile(localFile)
 }
 export default async function OpenAIServerComponent() {
-    const OpenAISourceDataRepository = serverContainer.get<SourceDataRepositoryOutputPort>(OPENAI.OPENAI_SOURCE_DATA_REPOSITORY);
+    const OpenAISourceDataRepository = serverContainer.get<SourceDataGatewayOutputPort>(OPENAI.OPENAI_SOURCE_DATA_REPOSITORY);
     const loggerFactory = serverContainer.get<(module: string) => Logger>(UTILS.LOGGER_FACTORY);
     const logger = loggerFactory("OpenAIServerComponent");
     // const listSourceDataDTO = await OpenAISourceDataRepository.list();
