+++ src/lib/core/dto/source-data-gateway-dto.ts
