+++ src/lib/infrastructure/client/config/ioc/client-ioc-symbols.ts
@@ -12,8 +12,6 @@ export const UTILS = {
     LOGGER_FACTORY: Symbol("LOGGER_FACTORY"),
 }
 export const REPOSITORY = {
-    BROWSER_SOURCE_DATA_REPOSITORY: Symbol("BROWSER_SOURCE_DATA_REPOSITORY"),
-    KERNEL_FILE_REPOSITORY: Symbol("KERNEL_FILE_REPOSITORY"),
 }
 
 export const GATEWAYS = {
@@ -21,6 +19,7 @@ export const GATEWAYS = {
     CONVERSATION_GATEWAY: Symbol("CONVERSATION_GATEWAY"),
     RESEARCH_CONTEXT_GATEWAY: Symbol("BROWSER_RESEARCH_CONTEXT_GATEWAY"),
     VECTOR_STORE_GATEWAY: Symbol("VECTOR_STORE_GATEWAY"),
+    SOURCE_DATA_GATEWAY: Symbol("SOURCE_DATA_GATEWAY"),
 }
 
 export const CONTROLLERS = {
@@ -33,9 +32,4 @@ export const CONTROLLERS = {
     LIST_RESEARCH_CONTEXTS_CONTROLLER: Symbol("LIST_RESEARCH_CONTEXTS_CONTROLLER"),
     LIST_SOURCE_DATA_CONTROLLER: Symbol("LIST_SOURCE_DATA_CONTROLLER"),
     SEND_MESSAGE_TO_CONVERSATION_CONTROLLER: Symbol("SEND_MESSAGE_TO_CONVERSATION_CONTROLLER"),
-}
-
-export const USECASE_FACTORY = {
-    CREATE_RESEARCH_CONTEXT: Symbol("CREATE_RESEARCH_CONTEXT"),
-    SEND_MESSAGE_TO_CONVERSATION: Symbol("SEND_MESSAGE_TO_CONVERSATION"),
 }
